requirements.txt
google-genai==1.12.1
python-dotenv==1.1.0
pylsp-mypy==0.7.0
